requirements.txt
asgiref==3.5.2
Django==4.1.1
django-ckeditor==6.5.1
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-ninja==0.19.1
django-restframework==0.0.1
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.2.1
idna==3.3
Pillow==9.2.0
pyasn1==0.4.8
pydantic==1.10.2
python-jose==3.3.0
rsa==4.9
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.2
